get readirq working for local interrupts

+++ defs.h
@@ -63,8 +63,8 @@ void popcli(void);
 
 // pic.c/bcmint.c
 void intinit(void);
-void intenable(uchar irq);
-uchar readirq(void);
+void intenable(u32 irq);
+u32 readirq(void);
 
 // proc.c
 void panic(char *fmt, ...) __attribute__((noreturn));
